+++ test/msw/handlers/private-messages/private-messages.handlers.ts
@@ -9,9 +9,7 @@ export const privateMessagesHandlers = {
         if (req.url.searchParams.get('a') !== '6') return res(ctx.status(404));
         return res(
           ctx.status(200),
-          ctx.text(
-            readHandlerMockFile('private-messages/send/send.success.html'),
-          ),
+          ctx.text(readHandlerMockFile('private-messages/send/success.html')),
         );
       }),
     ],
@@ -21,7 +19,7 @@ export const privateMessagesHandlers = {
         return res(
           ctx.status(200),
           ctx.text(
-            readHandlerMockFile('private-messages/send/send.invalid-user.html'),
+            readHandlerMockFile('private-messages/send/invalid-user.html'),
           ),
         );
       }),
@@ -31,8 +29,54 @@ export const privateMessagesHandlers = {
         if (req.url.searchParams.get('a') !== '6') return res(ctx.status(404));
         return res(
           ctx.status(200),
+          ctx.text(readHandlerMockFile('private-messages/send/unknown.html')),
+        );
+      }),
+    ],
+  },
+  replyOrForward: {
+    success: [
+      rest.get(`${forumConfig.FORUM_URL}pm`, (req, res, ctx) => {
+        if (
+          req.url.searchParams.get('a') !== '5' ||
+          !req.url.searchParams.get('reply')
+        )
+          return res(ctx.status(404));
+        return res(
+          ctx.status(200),
           ctx.text(
-            readHandlerMockFile('private-messages/send/send.unnknown.html'),
+            readHandlerMockFile(
+              'private-messages/reply-or-forward/success.html',
+            ),
+          ),
+        );
+      }),
+    ],
+    notFound: [
+      rest.get(`${forumConfig.FORUM_URL}pm`, (req, res, ctx) => {
+        if (
+          req.url.searchParams.get('a') !== '5' ||
+          !req.url.searchParams.get('reply')
+        )
+          return res(ctx.status(404));
+        return res(
+          ctx.status(200),
+          ctx.text(
+            readHandlerMockFile(
+              'private-messages/reply-or-forward/not-found.html',
+            ),
+          ),
+        );
+      }),
+    ],
+    unknown: [
+      rest.get(`${forumConfig.FORUM_URL}pm`, (req, res, ctx) => {
+        return res(
+          ctx.status(200),
+          ctx.text(
+            readHandlerMockFile(
+              'private-messages/reply-or-forward/unknown.html',
+            ),
           ),
         );
       }),